画面のファイルはpageフォルダに移動
