+++ frontend/src/components/page/AccountIntroduction.vue
