+++ frontend/src/components/page/AddKnowledge.vue
@@ -3,8 +3,8 @@ import { ref, onMounted } from 'vue';
 import axios from 'axios';
 import Cookies from 'js-cookie';
 
-import DropFile from './DropFile.vue'
-import Header from './Header.vue'
+import DropFile from '../DropFile.vue'
+import Header from '../Header.vue'
 
 const title = ref("");
 const knowledge = ref("");
