+++ frontend/src/components/page/EditProfile.vue
@@ -4,8 +4,8 @@ import { useRoute } from 'vue-router'
 import axios from 'axios';
 import Cookies from 'js-cookie';
 
-import DropFile from './DropFile.vue'
-import Header from './Header.vue'
+import DropFile from '../DropFile.vue'
+import Header from '../Header.vue'
 
 const route = useRoute();
