+++ frontend/src/components/page/EditRecord.vue
@@ -4,9 +4,9 @@ import axios from 'axios';
 import { useRoute } from 'vue-router'
 import Cookies from 'js-cookie';
 
-import DropFile from './DropFile.vue'
-import DatePicker from './DatePicker.vue'
-import Header from './Header.vue'
+import DropFile from '../DropFile.vue'
+import DatePicker from '../DatePicker.vue'
+import Header from '../Header.vue'
 
 const route = useRoute();
