+++ frontend/src/components/page/Setting.vue
@@ -4,7 +4,7 @@ import { useRouter } from 'vue-router'
 import axios from 'axios';
 import Cookies from 'js-cookie'
 
-import Header from './Header.vue'
+import Header from '../Header.vue'
 
 const router = useRouter();
