+++ frontend/src/components/page/Signup.vue
@@ -4,7 +4,7 @@ import { useRouter } from 'vue-router'
 import Cookies from 'js-cookie'
 import axios from 'axios'
 
-import Header from './Header.vue'
+import Header from '../Header.vue'
 
 const router = useRouter();
