+++ frontend/src/router/index.js
@@ -1,24 +1,24 @@
 import { createRouter, createWebHistory } from 'vue-router';
-import Home from '/src/components/Home.vue';
-import AccountInteroduction from '/src/components/AccountIntroduction.vue';
-import Signup from '/src/components/Signup.vue';
-import Login from '/src/components/Login.vue';
-import AddRecord from '/src/components/AddRecord.vue';
-import KnowledgeList from '/src/components/KnowledgeList.vue';
-import AddKnowledge from '/src/components/AddKnowledge.vue';
-import EditKnowledge from '/src/components/EditKnowledge.vue';
-import KnowledgeDetail from '/src/components/KnowledgeDetail.vue';
-import OtherRecordList from '/src/components/OtherRecordList.vue';
-import RecordDetail from '/src/components/RecordDetail.vue';
-import EditRecord from '/src/components/EditRecord.vue';
-import UserProfile from '/src/components/UserProfile.vue';
-import EditProfile from '/src/components/EditProfile.vue';
-import Setting from '/src/components/Setting.vue';
-import PasswordEdit from '/src/components/PasswordEdit.vue';
-import MailAddressEdit from '/src/components/MailAddressEdit.vue';
-import Contact from '/src/components/Contact.vue';
-import ContactList from '/src/components/ContactList.vue';
-import ContactDetail from '/src/components/ContactDetail.vue';
+import Home from '/src/components/page/Home.vue';
+import AccountInteroduction from '/src/components/page/AccountIntroduction.vue';
+import Signup from '/src/components/page/Signup.vue';
+import Login from '/src/components/page/Login.vue';
+import AddRecord from '/src/components/page/AddRecord.vue';
+import KnowledgeList from '/src/components/page/KnowledgeList.vue';
+import AddKnowledge from '/src/components/page/AddKnowledge.vue';
+import EditKnowledge from '/src/components/page/EditKnowledge.vue';
+import KnowledgeDetail from '/src/components/page/KnowledgeDetail.vue';
+import OtherRecordList from '/src/components/page/OtherRecordList.vue';
+import RecordDetail from '/src/components/page/RecordDetail.vue';
+import EditRecord from '/src/components/page/EditRecord.vue';
+import UserProfile from '/src/components/page/UserProfile.vue';
+import EditProfile from '/src/components/page/EditProfile.vue';
+import Setting from '/src/components/page/Setting.vue';
+import PasswordEdit from '/src/components/page/PasswordEdit.vue';
+import MailAddressEdit from '/src/components/page/MailAddressEdit.vue';
+import Contact from '/src/components/page/Contact.vue';
+import ContactList from '/src/components/page/ContactList.vue';
+import ContactDetail from '/src/components/page/ContactDetail.vue';
 
 const routes = [
   {
